move textBasedAnswerBackgroundTracker
